replace dialog with card component
